texturas, norma, leaks, sanitizer, movimiento

+++ includes/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: alvicina <alvicina@student.42urduliz.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/16 10:05:05 by alvicina          #+#    #+#             */
-/*   Updated: 2024/03/05 18:58:29 by alvicina         ###   ########.fr       */
+/*   Updated: 2024/03/06 19:10:49 by alvicina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -144,7 +144,7 @@ void	turn_pov(t_data *data, t_direction dir);
 // map_walls.c
 int		is_next_step_a_wall(t_data *data, double dir_rad);
 double	get_distance_to_wall(t_data *data, double x, double y, double dir_deg, int depth);
-int		is_coord_a_wall(t_data *data, int x, int y);
+int		is_coord_a_wall(t_data *data, double x, double y);
 
 // map_render.c
 void	render_map(t_data *data);
